fix: consolidate plex response structs
