+++ backend/internal/modals/PlexResponses.go
@@ -74,3 +74,17 @@ type PlexDirectoryItem struct {
 		Path string `xml:"path,attr"`
 	} `xml:"Location"`
 }
+
+type PlexPhotosResponse struct {
+	XMLName xml.Name        `xml:"MediaContainer"`
+	Size    int             `xml:"size,attr"`
+	Photos  []PlexPhotoItem `xml:"Photo"`
+}
+
+type PlexPhotoItem struct {
+	Key       string `xml:"key,attr"`
+	RatingKey string `xml:"ratingKey,attr"`
+	Thumb     string `xml:"thumb,attr"`
+	Selected  int    `xml:"selected,attr"`
+	Provider  string `xml:"provider,attr,omitempty"`
+}
